improve organization of utils modules
