+++ apis/python/src/tiledbsc/v1/util.py
@@ -1,4 +1,3 @@
-import re
 import time
 from typing import TypeVar
 
@@ -37,89 +36,7 @@ def is_local_path(path: str) -> bool:
     return True
 
 
-def is_tiledb_creation_uri(uri: str) -> bool:
-    return bool(re.match("^tiledb://.*s3://.*$", uri))
-
-
-def _to_tiledb_supported_dtype(dtype: np.dtype) -> np.dtype:
-    """A handful of types are cast into the TileDB type system."""
-    # TileDB has no float16 -- cast up to float32
-    if dtype == np.dtype("float16"):
-        return np.dtype("float32")
-    return dtype
-
-
-def _to_tiledb_supported_array_type(x: T) -> T:
-    """
-    Converts datatypes unrepresentable by TileDB into datatypes it can represent.
-    E.g., categorical strings -> string.
-
-    See also https://docs.scipy.org/doc/numpy-1.10.1/reference/arrays.dtypes.html
-
-    Preferentially converts to the underlying primitive type, as TileDB does not
-    support most complex types. NOTE: this does not support `datetime64` conversion.
-
-    Categoricals are a special case. If the underlying categorical type is a
-    primitive, convert to that. If the array contains NA/NaN (i.e. not in the
-    category, code == -1), raise error unless it is a float or string.
-    """
-    if isinstance(x, pd.DataFrame):
-        return pd.DataFrame.from_dict(
-            {k: _to_tiledb_supported_array_type(v) for k, v in x.items()}
-        )
-
-    # If a Pandas categorical, use the type of the underlying category.
-    # If the array contains NaN/NA, and the primitive is unable to represent
-    # a reasonable facsimile, i.e. not string or float, raise.
-    if pd.api.types.is_categorical_dtype(x.dtype):
-        categories = x.cat.categories
-        cat_dtype = categories.dtype
-        if cat_dtype.kind in ["f", "u", "i"]:
-            if x.hasnans and cat_dtype.kind == "i":
-                raise ValueError(
-                    "Categorical array contains NaN -- unable to convert to TileDB array."
-                )
-
-            return x.astype(_to_tiledb_supported_dtype(cat_dtype))
-
-        # Into the weirdness. See if Pandas can help with edge cases.
-        inferred = pd.api.types.infer_dtype(categories)
-        if inferred == "boolean":
-            if x.hasnans:
-                raise ValueError(
-                    "Categorical array contains NaN -- unable to convert to TileDB array."
-                )
-            return x.astype(bool)
-
-        if inferred == "string":
-            return x.astype(str)
-
-        if inferred == "bytes":
-            if x.hasnans:
-                raise ValueError(
-                    "Categorical array contains NaN -- unable to convert to TileDB array."
-                )
-            return x.astype(bytes)
-
-        return x.astype("O")
-
-    target_dtype = _to_tiledb_supported_dtype(x.dtype)
-    return x if target_dtype == x.dtype else x.astype(target_dtype)
-
-
-def _ascii_to_unicode_pandas_readback(df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Implements the 'decode on read' part of our ASCII/Unicode logic
-    """
-    # TODO: COMMENT/LINK HEAVILY
-    for k in df:
-        dfk = df[k]
-        if len(dfk) > 0 and type(dfk.iat[0]) == bytes:
-            df[k] = dfk.map(lambda e: e.decode())
-    return df
-
-
-def _find_csr_chunk_size(
+def find_csr_chunk_size(
     mat: sp.csr_matrix,
     start_row_index: int,
     goal_chunk_nnz: int,
@@ -146,7 +63,7 @@ def _find_csr_chunk_size(
 
 # This function is very similar to _find_csr_chunk_size. The code is largely repeated, and this is
 # intentional.  Here we err on the side of increased readability, at the expense of line-count.
-def _find_csc_chunk_size(
+def find_csc_chunk_size(
     mat: sp.csc_matrix,
     start_col_index: int,
     goal_chunk_nnz: int,
